requirements.txt

Package                       Version
----------------------------- -----------
amqp                          5.2.0
asgiref                       3.8.1
billiard                      4.2.1
celery                        5.4.0
certifi                       2024.8.30
charset-normalizer            3.3.2
click                         8.1.7
click-didyoumean              0.3.1
click-plugins                 1.1.1
click-repl                    0.3.0
colorama                      0.4.6
Django                        5.1.1
djangorestframework           3.15.2
djangorestframework-simplejwt 5.3.1
idna                          3.10
kombu                         5.4.2
pip                           23.1.2
prompt_toolkit                3.0.48
psycopg2                      2.9.9
PyJWT                         2.9.0
python-dateutil               2.9.0.post0
python-decouple               3.8
requests                      2.32.3
setuptools                    65.5.0
six                           1.16.0
sqlparse                      0.5.1
tzdata                        2024.2
urllib3                       2.2.3
vine                          5.1.0
wcwidth                       0.2.13

